% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createEstimateList}
\alias{createEstimateList}
\title{Create a List of synthdid_estimate Objects from a synthdid_estimate_multi Object}
\usage{
createEstimateList(est_did, k_plot = NULL)
}
\arguments{
\item{est_did}{A \code{synthdid_estimate_multi} object containing multiple estimates and associated
attributes such as weights, setup information, options, and method.}
}
\value{
A list of \code{synthdid_estimate} objects, where each object corresponds to an estimate
from the input \code{synthdid_estimate_multi} object. The returned list has the same length as the
number of estimates in the input object. Each \code{synthdid_estimate} object includes a two-dimensional
\code{Y} matrix and inherits other attributes from the input multi estimate object.
}
\description{
This function transforms a \code{synthdid_estimate_multi} object, which contains multiple estimates
for synthetic difference-in-differences analysis, into a list of \code{synthdid_estimate} objects.
Each \code{synthdid_estimate} object in the list represents one of the estimates from the
\code{synthdid_estimate_multi} object. The function properly handles the transformation of the \code{Y}
array from a three-dimensional array in the multi estimate object to a two-dimensional matrix
for each individual estimate, ensuring that each \code{synthdid_estimate} object has its unique \code{Y}
matrix. Other attributes such as \code{estimator}, \code{weights}, \code{setup}, \code{opts}, and \code{method} are
also appropriately copied and set for each individual estimate object.
}
\examples{
\donttest{
# Assuming est_did is a synthdid_estimate_multi object with K estimates
est_list <- createEstimateList(est_did)
# Now est_list contains K synthdid_estimate objects
}

}
