% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{format.synthdid_estimate_multi}
\alias{format.synthdid_estimate_multi}
\title{Format a synthdid_multi object}
\usage{
\method{format}{synthdid_estimate_multi}(x, ...)
}
\arguments{
\item{x}{The object to format}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Format a synthdid_multi object
}
