% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov_multivar.R
\name{jackknife_multi}
\alias{jackknife_multi}
\title{Jackknife standard error of function \code{theta} at samples \code{x}.}
\usage{
jackknife_multi(x, theta, K)
}
\arguments{
\item{x}{vector of samples}

\item{theta}{a function which returns a scalar estimate}
}
\description{
Jackknife standard error of function \code{theta} at samples \code{x}.
}
\keyword{internal}
