% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{panel.array}
\alias{panel.array}
\title{Convert a long (balanced) panel to a array}
\usage{
panel.array(
  panel,
  unit = 1,
  time = 2,
  category = 3,
  outcome = 4,
  treatment = 5,
  treated.last = TRUE
)
}
\arguments{
\item{panel}{A data.frame with columns consisting of units, time, catgory, outcome, and treatment indicator.}

\item{unit}{The column number/name corresponding to the unit identifier. Default is 1.}

\item{time}{The column number/name corresponding to the time identifier. Default is 2.}

\item{category}{The column number/name corresponding to the outcome identifier. Default is 3.}

\item{outcome}{The column number/name corresponding to the outcome identifier. Default is 4.}

\item{treatment}{The column number/name corresponding to the treatment status. Default is 5.}

\item{treated.last}{Should we sort the rows of Y and W so treated units are last. If FALSE, sort by unit number/name. Default is TRUE.}
}
\value{
A list with entries \code{Y}: the data array, \code{N0}: the number of control units, \code{T0}:
the number of time periods before treatment, \code{W}: the arrray of treatment indicators.
}
\description{
Converts a data set in panel form with multivariate outcomes to matrix format required by synthdid estimators.
A typical long panel date set looks like [unit, time, category, outcome, treatment]. Synthdid
requires a balanced panel with simultaneous adoption of treatment: each unit must be observed
at all times, and all treated units must begin treatment simultaneosly. This function
creates num.units x num.time.periods matrices Y and W of outcomes and treatment indicators.
A slice for each catgory our multivariate outcome.
In these array, columns are sorted by time, and by default (when treated.last=TRUE),
rows for control units appear before those of treated units.
}
\examples{
\donttest{
}

}
