% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.synthdid_estimate_multi}
\alias{plot.synthdid_estimate_multi}
\title{Plot a synthdid multi object}
\usage{
\method{plot}{synthdid_estimate}(x, which_plot = NULL, ...)
}
\arguments{
\item{x}{The object to plot}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Plot a synthdid multi object
}
