% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.synthdid_estimate_multi}
\alias{print.synthdid_estimate_multi}
\title{Print a synthdid_multi object}
\usage{
\method{print}{synthdid_estimate}(x, ...)
}
\arguments{
\item{x}{The object to print}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a synthdid_multi object
}
