% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid-package.R
\docType{package}
\name{propsdid-package}
\alias{propsdid}
\alias{propsdid-package}
\title{propsdid: Estimating Treatment Effects on Proportions with Synthetic Controls}
\description{
The \code{propsdid} package provides methods for estimating treatment effects on
\emph{proportional outcomes} using extensions of the Synthetic Control (SC) and
Synthetic Difference‑in‑Differences (SDID) estimators.

The package implements Synthetic Control  and Synthetic
Difference‑in‑Differences ewith common-weights stimators introduced in:

Bogatyrev, K., & Stoetzer, L.
\emph{Estimating treatment effects on proportions with synthetic controls.}
OSF Preprints. https://osf.io/preprints/osf/brhd3

It extends the original \code{synthdid} package that estimates treatment effects for
\emph{single‑outcome panels}. This package includes an extension to settings
where \strong{multiple proportional outcomes}.

The method estimates a common set of unit and time weights across all
outcomes, ensuring that the reconstructed counterfactual respects the
compositional structure of the data, and treatment effects sum to zero.
\subsection{Key Features}{
\itemize{
\item Synthetic Difference‑in‑Differences with \strong{common weights}
\item Synthetic Control with \strong{common weights}
}
}

\subsection{Links}{
\itemize{
\item GitHub repository: https://github.com/lstoetze/propsdid
\item Paper preprint: https://osf.io/preprints/osf/brhd3
}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/synth-inference/synthdid}
  \item Report bugs at \url{https://github.com/synth-inference/synthdid/issues}
}

}
\author{
\strong{Maintainer}: Lukas Stoetzer \email{lukas.stoetzer@uni-wh.de}

Authors:
\itemize{
  \item Konstantin Bogatyrev \email{konstantin.bogatyrev@phd.unibocconi.it}
}

}
\keyword{internal}
