% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthdid.R
\name{sc_estimate}
\alias{sc_estimate}
\title{Estimates Synthetic Control models with common weights for multivariate
proportional outcomes.}
\usage{
sc_estimate(Y, N0, T0, method = "did", porp_dat = TRUE, ...)
}
\arguments{
\item{Y}{the observation matrix.}

\item{N0}{the number of control units. Rows 1-N0 of Y correspond to the control units.}

\item{T0}{the number of pre-treatment time steps. Columns 1-T0 of Y correspond to pre-treatment time steps.}

\item{method}{specifies what method should be used: sdid (synthetic difference in difference), sc (synthetic control), or did (difference in difference)}

\item{...}{additional options for synthdid_estimate}

\item{prop_dat}{specifies if outcome is multivariate (TRUE default) or univariate (FALSE)}
}
\value{
an object like that returned by synthdid_estimate
}
\description{
This function implements the common‑weights Synthetic Control estimator
introduced in \emph{Estimating Treatment Effects on Proportions with Synthetic
Controls} (Stoetzer & Bogatyrev). It applies the standard SC setup but
enforces a shared set of unit and time weights across all outcomes, ensuring
that the counterfactual respects the compositional structure of the data.
}
