% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.synthdid_estimate_multi}
\alias{summary.synthdid_estimate_multi}
\title{Summarize a synthdid_multi object}
\usage{
\method{summary}{synthdid_estimate_multi}(object, ...)
}
\arguments{
\item{object}{The object to summarize}

\item{...}{Additional arguments (currently ignored).}

\item{weight.digits}{The number of digits to use when displaying weights (omega, lambda)}

\item{fast}{Be fast but less accurate, e.g. jackknife instead of bootstrap se estimate}
}
\description{
Summarize a synthdid_multi object
}
